feat: Implement tasks/list client request
This adds the core to the client for invoking the `tasks/list` RPC request.

As part of this, the types from the a2a spec have been regenerated from the last commit which still published a discrete a2a.json artifact.

It did involve a small update to the `ParamsOf` type to be able to cope with optional parameters as well as mandatory ones.